Extract shared corpus prep; add project-scoped storage design note
The Pride and Prejudice prep was copy-pasted into gpt.py, gpt_sweep.py and
gpt-sweep/experiment.py. Pull download_pride_and_prejudice/prepare_data into
one importable experiment/corpus.py so experiments share a single symbol —
the seed of cross-experiment reuse, since the memo key is already
content-addressed and name-independent.

Also add notes/storage-design.md, capturing the agreed direction: one
volume + one memo store per project (not per experiment), with the project
id derived from the nearest pyproject [project].name (workspace member in a
monorepo) and a MINI_PROJECT override. Documents the required follow-ups
(record tagging + CLI filtering, name->resource decoupling, content-addressed
dataset paths as the read-only replacement) and the HF-from-volume remote
publish step.

Note: gpt.py / gpt_sweep.py published HTML is now stale vs source — re-run
to refresh (needs GPU/Modal).
